better organizing code related to syntax & color schemes
